clinic_ai_backend/requirements.txt
fastapi
uvicorn
pydantic
pymongo
python-dotenv
pytest
ruff
openai-whisper
